refactor(Lotto): LottoCalculator 와 Lotto 명확히 분리
