+++ src/main/java/step3/view/InputView.java
@@ -1,7 +1,9 @@
 package step3.view;
 
+import step3.model.Lotto;
 import step3.model.LottoCalculator;
 import step3.model.LottoGenerator;
+import step3.model.LottoNumber;
 
 import java.util.Scanner;
 
@@ -14,13 +16,15 @@ public class InputView {
         lottoGenerator.setPurchasePrice(scanner.nextLine());
     }
 
-    public static void inputLastWeekLottoNumbers(LottoCalculator calculator) {
+    public static void inputLastWeekLottoNumbers(Lotto lotto) {
         System.out.println(LAST_LOTTO_NUMBERS_INPUT_MESSAGE);
-        calculator.setLastWeekLottoNumbers(scanner.nextLine());
+        String[] lastweekNumbers = lotto.validateLastWeekWinner(scanner.nextLine());
+
+        lotto = inputLastWeekBonusNumber(lastweekNumbers);
     }
 
-    public static void inputLastWeekBonusNumber(LottoCalculator calculator) {
+    public static Lotto inputLastWeekBonusNumber(String[] lastWeekLottos) {
         System.out.println(LAST_BONUS_NUMBER_INPUT_MESSAGE);
-        calculator.setLastWeekBonusNumber(scanner.nextLine());
+        return new Lotto(lastWeekLottos, Integer.parseInt(scanner.nextLine()));
     }
 }
